refactor JS to object-oriented design
This refactoring was a big pain, and I'm not super sure that it was
worth it, but it was starting to bug me how many variables were being
closed over and shared among the various functions (like
selectedLanguages, paused, showingForm, etc.). By moving to an object-
oriented design, those variables now have a more proper home, as
instance variables of an instance of the LanguageSelector class.