remove unneccessary `js: true` options.
I believe that there is some purpose for these, but I don't know what it
is. Our spec_helper currently has the default driver and the javascript
driver as the same thing (poltergeist), so maybe that has something to
do with why `js: true` doesn't seem to have any effect.