Update Ruby gems (bundle update)
I am primarily motivated to do this because GitHub actually warned me
that there is/was a security vulnerability in `rubyzip`, which is one of
the gems being upgraded with this PR. Fortunately, I don't think [that
that particular security vulnerability][1] had any substantive impact of
the security of this project, since that vulnerability is related to
using `rubyzip` to handle zip files uploaded to a server by potentially
malicious actors, for which purpose that gem was never utilized in this
project. Nevertheless, it's a good idea to update that gem (`rubyzip`)
and all gems in general.

[1]: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-5946