Use headless Chrome instead of poltergeist to run feature specs
I think that PhantomJS is on the way out, and that Headless Chrome is
going to be the future of testing. This change aims to keep this project
up-to-date with the latest testing tools, which I believe to be headless
Chrome, not PhantomJS/poltergeist. (This is not to detract from
PhantomJS/poltergeist -- they have been incredibly useful libraries to
me!)

It's nice, as well, to see that the changes to the Gemfile are primarily
deletions, i.e. we are dropping our dependency on a few libraries. Fewer
dependencies is always better, all other things being equal. :)